Move files and remove mocked beacon
